+++ main.go
@@ -44,7 +44,7 @@ import (
 )
 
 // Version 应用版本号
-var Version = "3.87.0-rc2"
+var Version = "3.88.0-rc1"
 
 // capitalDataSourceAdapter 资金數據源适配器
 type capitalDataSourceAdapter struct {